+++ app/src/main/kotlin/me/xizzhu/android/joshua/translations/TranslationListView.kt
@@ -38,6 +38,8 @@ interface TranslationView : BaseSettingsView {
 
     fun onTranslationsLoadingFailed(forceRefresh: Boolean)
 
+    fun onNoTranslationsAvailable()
+
     fun onTranslationsUpdated(translations: List<BaseItem>)
 
     fun onTranslationDownloadStarted()
@@ -125,9 +127,22 @@ class TranslationListView : BaseRecyclerView, TranslationView {
     }
 
     override fun onTranslationsLoadingFailed(forceRefresh: Boolean) {
-        DialogHelper.showDialog(context, true, R.string.dialog_load_translation_list_error,
+        DialogHelper.showDialog(context, false, R.string.dialog_load_translation_list_error,
                 DialogInterface.OnClickListener { _, _ ->
                     presenter.loadTranslationList(forceRefresh)
+                },
+                DialogInterface.OnClickListener { _, _ ->
+                    presenter.finish()
+                })
+    }
+
+    override fun onNoTranslationsAvailable() {
+        DialogHelper.showDialog(context, false, R.string.dialog_empty_translation_list_error,
+                DialogInterface.OnClickListener { _, _ ->
+                    presenter.loadTranslationList(true)
+                },
+                DialogInterface.OnClickListener { _, _ ->
+                    presenter.finish()
                 })
     }